+++ test_autofit/analysis/test_relation.py
@@ -1,0 +1,2 @@
+def test_free_parameter_as_model():
+    pass
